+++ TidalPy/tides/eccentricity_funcs/orderl3.py
@@ -6,7 +6,7 @@ from ...utilities.performance.numba import njit
 from ...utilities.types import FloatArray
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc2(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^2 for order-l = 3
     Parameters
@@ -54,7 +54,7 @@ def eccentricity_funcs_trunc2(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc4(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^4 for order-l = 3
     Parameters
@@ -111,7 +111,7 @@ def eccentricity_funcs_trunc4(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc6(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^6 for order-l = 3
     Parameters
@@ -175,7 +175,7 @@ def eccentricity_funcs_trunc6(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc8(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^8 for order-l = 3
     Parameters
@@ -248,7 +248,7 @@ def eccentricity_funcs_trunc8(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc10(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^10 for order-l = 3
     Parameters
@@ -330,7 +330,7 @@ def eccentricity_funcs_trunc10(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc12(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^12 for order-l = 3
     Parameters
@@ -421,7 +421,7 @@ def eccentricity_funcs_trunc12(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc14(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^14 for order-l = 3
     Parameters
@@ -521,7 +521,7 @@ def eccentricity_funcs_trunc14(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc16(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^16 for order-l = 3
     Parameters
@@ -630,7 +630,7 @@ def eccentricity_funcs_trunc16(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc18(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^18 for order-l = 3
     Parameters
@@ -748,7 +748,7 @@ def eccentricity_funcs_trunc18(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc20(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^20 for order-l = 3
     Parameters