+++ TidalPy/tides/eccentricity_funcs/orderl6.py
@@ -2,11 +2,11 @@
 """
 
 from . import EccenOutput
-from ...utilities.performance.numba import njit
+from ...utilities.performance.numba import njit, nbDict, float64, int16, prange
 from ...utilities.types import FloatArray
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc2(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^2 for order-l = 6
     Parameters
@@ -71,7 +71,7 @@ def eccentricity_funcs_trunc2(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc4(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^4 for order-l = 6
     Parameters
@@ -149,7 +149,7 @@ def eccentricity_funcs_trunc4(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc6(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^6 for order-l = 6
     Parameters
@@ -244,7 +244,7 @@ def eccentricity_funcs_trunc6(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc8(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^8 for order-l = 6
     Parameters
@@ -352,7 +352,7 @@ def eccentricity_funcs_trunc8(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc10(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^10 for order-l = 6
     Parameters
@@ -475,7 +475,7 @@ def eccentricity_funcs_trunc10(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc12(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^12 for order-l = 6
     Parameters
@@ -611,7 +611,7 @@ def eccentricity_funcs_trunc12(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc14(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^14 for order-l = 6
     Parameters
@@ -762,7 +762,7 @@ def eccentricity_funcs_trunc14(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc16(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^16 for order-l = 6
     Parameters
@@ -928,7 +928,7 @@ def eccentricity_funcs_trunc16(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc18(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^18 for order-l = 6
     Parameters
@@ -1109,7 +1109,7 @@ def eccentricity_funcs_trunc18(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc20(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^20 for order-l = 6
     Parameters