+++ TidalPy/tides/eccentricity_funcs/orderl7.py
@@ -6,7 +6,7 @@ from ...utilities.performance.numba import njit
 from ...utilities.types import FloatArray
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc2(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^2 for order-l = 7
     Parameters
@@ -74,7 +74,7 @@ def eccentricity_funcs_trunc2(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc4(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^4 for order-l = 7
     Parameters
@@ -161,7 +161,7 @@ def eccentricity_funcs_trunc4(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc6(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^6 for order-l = 7
     Parameters
@@ -263,7 +263,7 @@ def eccentricity_funcs_trunc6(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc8(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^8 for order-l = 7
     Parameters
@@ -384,7 +384,7 @@ def eccentricity_funcs_trunc8(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc10(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^10 for order-l = 7
     Parameters
@@ -520,7 +520,7 @@ def eccentricity_funcs_trunc10(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc12(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^12 for order-l = 7
     Parameters
@@ -673,7 +673,7 @@ def eccentricity_funcs_trunc12(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc14(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^14 for order-l = 7
     Parameters
@@ -841,7 +841,7 @@ def eccentricity_funcs_trunc14(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc16(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^16 for order-l = 7
     Parameters
@@ -1026,7 +1026,7 @@ def eccentricity_funcs_trunc16(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc18(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^18 for order-l = 7
     Parameters
@@ -1228,7 +1228,7 @@ def eccentricity_funcs_trunc18(eccentricity: FloatArray) -> EccenOutput:
     return eccentricity_results_bymode
 
 
-@njit(cacheable=True)
+@njit(cacheable=True, parallel=True)
 def eccentricity_funcs_trunc20(eccentricity: FloatArray) -> EccenOutput:
     """ Calculates the eccentricity functions (by mode) truncated to e^20 for order-l = 7
     Parameters